refactor auth to jwt
